requirements.txt
Django==1.8.1
django-mptt==0.7.3
